Move String utils to separate file
