+++ OpenTerm/Controller/TerminalViewController.swift
@@ -15,21 +15,6 @@ import PanelKit
 import StoreKit
 import MobileCoreServices
 
-extension String {
-	func toCString() -> UnsafePointer<Int8>? {
-		let nsSelf: NSString = self as NSString
-		return nsSelf.cString(using: String.Encoding.utf8.rawValue)
-	}
-}
-
-extension String {
-
-	var utf8CString: UnsafeMutablePointer<Int8> {
-		return UnsafeMutablePointer(mutating: (self as NSString).utf8String!)
-	}
-
-}
-
 class TerminalViewController: UIViewController {
 
 	let terminalView: TerminalView
